# proto-gen/openapi/api_v2.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api_v2.proto",
    "version": "1.0"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v2/samplingStrategy": {
      "post": {
        "operationId": "GetSamplingStrategy",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/api_v2SamplingStrategyResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api_v2SamplingStrategyParameters"
            }
          }
        ],
        "tags": [
          "SamplingManager"
        ]
      }
    },
    "/api/v2/spans": {
      "post": {
        "operationId": "PostSpans",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/api_v2PostSpansResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api_v2PostSpansRequest"
            }
          }
        ],
        "tags": [
          "CollectorService"
        ]
      }
    }
  },
  "definitions": {
    "api_v2Batch": {
      "type": "object",
      "properties": {
        "spans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2Span"
          }
        },
        "process": {
          "$ref": "#/definitions/api_v2Process"
        }
      }
    },
    "api_v2KeyValue": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "v_type": {
          "$ref": "#/definitions/api_v2ValueType"
        },
        "v_str": {
          "type": "string"
        },
        "v_bool": {
          "type": "boolean",
          "format": "boolean"
        },
        "v_int64": {
          "type": "string",
          "format": "int64"
        },
        "v_float64": {
          "type": "number",
          "format": "double"
        },
        "v_binary": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "api_v2Log": {
      "type": "object",
      "properties": {
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2KeyValue"
          }
        }
      }
    },
    "api_v2OperationSamplingStrategy": {
      "type": "object",
      "properties": {
        "operation": {
          "type": "string"
        },
        "probabilisticSampling": {
          "$ref": "#/definitions/api_v2ProbabilisticSamplingStrategy"
        }
      }
    },
    "api_v2PerOperationSamplingStrategies": {
      "type": "object",
      "properties": {
        "defaultSamplingProbability": {
          "type": "number",
          "format": "double"
        },
        "defaultLowerBoundTracesPerSecond": {
          "type": "number",
          "format": "double"
        },
        "perOperationStrategies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2OperationSamplingStrategy"
          }
        },
        "defaultUpperBoundTracesPerSecond": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "api_v2PostSpansRequest": {
      "type": "object",
      "properties": {
        "batch": {
          "$ref": "#/definitions/api_v2Batch"
        }
      }
    },
    "api_v2PostSpansResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "api_v2ProbabilisticSamplingStrategy": {
      "type": "object",
      "properties": {
        "samplingRate": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "api_v2Process": {
      "type": "object",
      "properties": {
        "service_name": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2KeyValue"
          }
        }
      }
    },
    "api_v2RateLimitingSamplingStrategy": {
      "type": "object",
      "properties": {
        "maxTracesPerSecond": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "api_v2SamplingStrategyParameters": {
      "type": "object",
      "properties": {
        "serviceName": {
          "type": "string"
        }
      }
    },
    "api_v2SamplingStrategyResponse": {
      "type": "object",
      "properties": {
        "strategyType": {
          "$ref": "#/definitions/api_v2SamplingStrategyType"
        },
        "probabilisticSampling": {
          "$ref": "#/definitions/api_v2ProbabilisticSamplingStrategy"
        },
        "rateLimitingSampling": {
          "$ref": "#/definitions/api_v2RateLimitingSamplingStrategy"
        },
        "operationSampling": {
          "$ref": "#/definitions/api_v2PerOperationSamplingStrategies"
        }
      }
    },
    "api_v2SamplingStrategyType": {
      "type": "string",
      "enum": [
        "PROBABILISTIC",
        "RATE_LIMITING"
      ],
      "default": "PROBABILISTIC"
    },
    "api_v2Span": {
      "type": "object",
      "properties": {
        "trace_id": {
          "type": "string",
          "format": "byte"
        },
        "span_id": {
          "type": "string",
          "format": "byte"
        },
        "operation_name": {
          "type": "string"
        },
        "references": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2SpanRef"
          }
        },
        "flags": {
          "type": "integer",
          "format": "int64"
        },
        "start_time": {
          "type": "string",
          "format": "date-time"
        },
        "duration": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2KeyValue"
          }
        },
        "logs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v2Log"
          }
        },
        "process": {
          "$ref": "#/definitions/api_v2Process"
        },
        "process_id": {
          "type": "string"
        },
        "warnings": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "api_v2SpanRef": {
      "type": "object",
      "properties": {
        "trace_id": {
          "type": "string",
          "format": "byte"
        },
        "span_id": {
          "type": "string",
          "format": "byte"
        },
        "ref_type": {
          "$ref": "#/definitions/api_v2SpanRefType"
        }
      }
    },
    "api_v2SpanRefType": {
      "type": "string",
      "enum": [
        "CHILD_OF",
        "FOLLOWS_FROM"
      ],
      "default": "CHILD_OF"
    },
    "api_v2ValueType": {
      "type": "string",
      "enum": [
        "STRING",
        "BOOL",
        "INT64",
        "FLOAT64",
        "BINARY"
      ],
      "default": "STRING"
    }
  },
  "externalDocs": {
    "description": "Jaeger API",
    "url": "https://github.com/jaegertracing/jaeger"
  }
}